Remove MoveManager interface as it is not needed
